+++ crates/runtime/src/prelude.rs
@@ -4,8 +4,8 @@
 pub use crate::{
     make_ptr, make_ptr_mut, runtime_error, type_error, type_error_with_slice, BinaryOp, CallArgs,
     CallContext, DisplayContext, IsIterable, KCell, KIterator, KIteratorOutput, KList, KMap,
-    KNativeFunction, KNumber, KObject, KRange, KString, KTuple, KValue, KotoCopy, KotoFile,
-    KotoFunction, KotoHasher, KotoIterator, KotoLookup, KotoObject, KotoRead, KotoSend, KotoSync,
+    KNativeFunction, KNumber, KObject, KRange, KString, KTuple, KValue, KotoCopy, KotoEntries,
+    KotoFile, KotoFunction, KotoHasher, KotoIterator, KotoObject, KotoRead, KotoSend, KotoSync,
     KotoType, KotoVm, KotoVmSettings, KotoWrite, MetaKey, MetaMap, MethodContext, UnaryOp,
     ValueKey, ValueMap, ValueVec,
 };